requirements.txt
numpy
pandas
matplotlib
xgboost
cupy-cuda12x
